Finish speech recognition tutorial
I did a bit of reading and learned about engine layouts web browsers have and what the differences are
More adding html via Javascript, really good to build upon that.
Another extremely fun and interesting tutorial.  In particular, using includes on the transcript of the input to call functions or log something is very useful.