Finish Follow Along Links tutorial
using `Element.getBoundingClientRect()` to add a style to elements.  In this case we added a highlight to all anchor tags

The Element.getBoundingClientRect() method returns the size of an element and its position relative to the viewport.
Using this we added the windows x + y position to the anchor tags top and left values to add a highlight with transform translate